Link /etc/machine-id to a real ID
It seems a reasonable approach to symlink /etc/machine-id to a 'real' ID
which will persist on the device and is immutable. In particular, this
may benefit read-only filesystems deployments where the file contents is
required (or expected) to be consistent across boots.

There is a concern about doing this because it will prevent systemd
from running units with ConditionFirstBoot=yes [1]
However, this must already be the case today because:

 * Debian's systemd pkg postinst hook runs systemd-machine-id-setup in
   the chroot which populates it with a 'valid' (generated) ID.
 * pi-gen writes the empty string to it regardless if it exists.
 * mmdebstrap writes the empty string to it if it exists [2].

Any one of these operations will have the same effect (as far as systemd
is concerned) as sym linking the file to
/sys/firmware/devicetree/base/serial-number

Reflect this change in the in-tree image layout rootfs overlays.
We can't assume this symlink would be something that all deployments
would like, so if a custom board or image overlay wants to do different,
it can.

[1] https://www.freedesktop.org/software/systemd/man/latest/machine-id.html
[2] https://manpages.debian.org/testing/mmdebstrap/mmdebstrap.1.en.html#cleanup